move is_numerical_scalar
